requirements.txt
boto3==1.35.36
botocore==1.35.36
chardet==5.2.0
jmespath==1.0.1
pillow==10.4.0
PyMuPDF==1.24.11
python-dateutil==2.9.0.post0
reportlab==4.2.5
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
